+++ Source/Custom Views/Shortcut Recorder/CTGradient.h
@@ -11,8 +11,8 @@
 
 typedef struct _CTGradientElement 
 	{
-	float red, green, blue, alpha;
-	float position;
+	CGFloat red, green, blue, alpha;
+	CGFloat position;
 	
 	struct _CTGradientElement *nextElement;
 	} CTGradientElement;
